Splitting NotificationManagerWrapper.scala into a number of smaller files
NotificationManagerWrapper.scala was a messy file with a bunch of related but different classes used for defining
how a displayed notification should look like. I moved them all to separate files.
